refactor(workstation): unify launch actions; ephemeral Launchpad tab
Share one action list between the empty-pool launcher and the `+` menu,
and rebrand the launcher as the Launchpad.

- Extract useWorkStationLaunchActions as the single source of truth for
  the eight launch actions (search / Explorer / Source Control / Terminal
  / browser / private browser / work items / projects). Both the start
  page tiles and the TabBarPlusMenu dropdown render from it, so their
  items and icons always match — the dropdown now also offers Explorer /
  Source Control / Terminal.
- Rebrand the `start` tab as "Launchpad" with the LayoutGrid icon (the
  same icon the chat panel uses); register LayoutGrid in the tab-icon map
  and title the launcher page "Launchpad".
- Make the Launchpad an ephemeral placeholder (useLaunchpadTab, renamed
  from useEnsureStartTab): seed it when the pool empties, drop it once a
  real tab exists — so closing the last tab brings it back and opening
  anything replaces it.

Pre-commit hook ran. Total eslint: 0, total circular: 0
