+++ src/modules/WorkStation/AppShell/StartPage/index.tsx
@@ -1,46 +1,20 @@
 /**
  * WorkStationStartPage
  *
- * Launcher shown when the WorkStation tab pool is empty (on app start, and
- * whenever the user closes the last tab). Lets the user pick what to open —
- * mirroring the unified `+` menu's actions (TabBarPlusMenu). No tab is
- * auto-opened; everything lazy-loads from here.
+ * Launchpad shown when the WorkStation tab pool holds only the launcher tab
+ * (on app start, and whenever the user closes the last real tab). Lets the
+ * user pick what to open — sharing its action list with the unified `+` menu
+ * (TabBarPlusMenu) via `useWorkStationLaunchActions` so the two always match.
  */
-import { useSetAtom } from "jotai";
-import {
-  Box,
-  FileSearch,
-  FolderTree,
-  GitBranch,
-  Globe,
-  ListTodo,
-  type LucideIcon,
-  ShieldOff,
-  SquareTerminal,
-} from "lucide-react";
-import React, { memo, useCallback, useMemo } from "react";
+import { type LucideIcon } from "lucide-react";
+import React, { memo } from "react";
 import { useTranslation } from "react-i18next";
 
 import KeyBadge from "@src/components/KeyBadge";
-import { getShortcutKeys } from "@src/config/keyboard/shortcutDisplay";
 import { SURFACE_TOKENS } from "@src/config/surfaceTokens";
-import { focusBrowserUrlBar } from "@src/modules/WorkStation/Browser/Panels/BrowserMainPane/components/WebUrlBar";
 import { EDITOR_TAB_CANVAS_BG_CLASS } from "@src/modules/WorkStation/shared/tokens";
-import { openEditorSpotlight } from "@src/scaffold/GlobalSpotlight/openSpotlight";
-import {
-  CODE_EDITOR_MAIN_TERMINAL_SESSION_ID,
-  STORY_ORG_SCOPE,
-  createExplorerTab,
-  createProjectDashboardTab,
-  createProjectWorkItemsIndexTab,
-  createSourceControlTab,
-  createTerminalTab,
-  dockFilterAtom,
-  openTab as openTabMutation,
-  requestNewBrowserSessionAtom,
-  workstationLayoutAtom,
-} from "@src/store/workstation";
-import type { WorkStationTab } from "@src/store/workstation/tabs";
+
+import { useWorkStationLaunchActions } from "../useWorkStationLaunchActions";
 
 interface StartActionTileProps {
   icon: LucideIcon;
@@ -68,95 +42,7 @@ StartActionTile.displayName = "StartActionTile";
 
 export const WorkStationStartPage: React.FC = memo(() => {
   const { t } = useTranslation("navigation");
-  const requestNewBrowserSession = useSetAtom(requestNewBrowserSessionAtom);
-  const setLayout = useSetAtom(workstationLayoutAtom);
-  const setDockFilter = useSetAtom(dockFilterAtom);
-
-  const openTabInMainPane = useCallback(
-    (tab: WorkStationTab) => {
-      setLayout((prev) => {
-        if (!prev?.mainPane) return prev;
-        return { ...prev, mainPane: openTabMutation(prev.mainPane, tab) };
-      });
-    },
-    [setLayout]
-  );
-
-  // The Browser host keeps its sessions in a separate store, so opening a
-  // browser tab means revealing that host (`dockFilter = "browser"`) rather
-  // than adding a `mainPane` tab. The pending new-session request is honored
-  // once the host is mounted.
-  const openBrowser = useCallback(
-    (isPrivate: boolean) => {
-      setDockFilter("browser");
-      requestNewBrowserSession(isPrivate ? { isPrivate: true } : {});
-      focusBrowserUrlBar();
-    },
-    [setDockFilter, requestNewBrowserSession]
-  );
-
-  const actions = useMemo<StartActionTileProps[]>(
-    () => [
-      {
-        icon: FileSearch,
-        label: t("workstation.plusMenu.searchFile"),
-        shortcut: "⌘P",
-        onClick: () => openEditorSpotlight(""),
-      },
-      {
-        icon: FolderTree,
-        label: t("workstation.startPage.explorer"),
-        shortcut: getShortcutKeys("open_file_folder_tab"),
-        onClick: () => openTabInMainPane(createExplorerTab()),
-      },
-      {
-        icon: GitBranch,
-        label: t("workstation.startPage.sourceControl"),
-        shortcut: getShortcutKeys("open_source_control_tab"),
-        onClick: () =>
-          openTabInMainPane(createSourceControlTab(0, { mode: "all-changes" })),
-      },
-      {
-        icon: SquareTerminal,
-        label: t("workstation.startPage.terminal"),
-        shortcut: getShortcutKeys("open_terminal_tab"),
-        onClick: () =>
-          openTabInMainPane(
-            createTerminalTab(
-              CODE_EDITOR_MAIN_TERMINAL_SESSION_ID,
-              t("workstation.startPage.terminal")
-            )
-          ),
-      },
-      {
-        icon: Globe,
-        label: t("workstation.plusMenu.newBrowserTab"),
-        onClick: () => openBrowser(false),
-      },
-      {
-        icon: ShieldOff,
-        label: t("workstation.plusMenu.newPrivateBrowserTab"),
-        onClick: () => openBrowser(true),
-      },
-      {
-        icon: ListTodo,
-        label: t("workstation.plusMenu.workItems"),
-        onClick: () =>
-          openTabInMainPane(
-            createProjectWorkItemsIndexTab({ orgScope: STORY_ORG_SCOPE.ALL })
-          ),
-      },
-      {
-        icon: Box,
-        label: t("workstation.plusMenu.projects"),
-        onClick: () =>
-          openTabInMainPane(
-            createProjectDashboardTab({ orgScope: STORY_ORG_SCOPE.ALL })
-          ),
-      },
-    ],
-    [openTabInMainPane, openBrowser, t]
-  );
+  const actions = useWorkStationLaunchActions();
 
   return (
     <div
@@ -164,14 +50,20 @@ export const WorkStationStartPage: React.FC = memo(() => {
     >
       <div className="w-full max-w-[560px]">
         <h1 className="mb-1 text-[20px] font-semibold text-text-1">
-          {t("workstation.startPage.title")}
+          {t("routes.launchpad")}
         </h1>
         <p className="mb-6 text-[13px] leading-snug text-text-3">
           {t("workstation.startPage.subtitle")}
         </p>
         <div className="grid grid-cols-2 gap-3 sm:grid-cols-3">
           {actions.map((action) => (
-            <StartActionTile key={action.label} {...action} />
+            <StartActionTile
+              key={action.id}
+              icon={action.icon}
+              label={action.label}
+              shortcut={action.shortcut}
+              onClick={action.onClick}
+            />
           ))}
         </div>
       </div>
