+++ src/modules/WorkStation/AppShell/index.tsx
@@ -34,7 +34,7 @@ import { useAppShellRepo } from "./hooks/useAppShellRepo";
 import { useAppShellSimulatorPanelSync } from "./hooks/useAppShellSimulatorPanelSync";
 import { useAppShellStationMode } from "./hooks/useAppShellStationMode";
 import { useAppShellStatusBar } from "./hooks/useAppShellStatusBar";
-import { useEnsureStartTab } from "./hooks/useEnsureStartTab";
+import { useLaunchpadTab } from "./hooks/useLaunchpadTab";
 
 interface AppShellProps {
   /** Whether WorkStation is currently visible (code view mode is active) */
@@ -76,9 +76,10 @@ const AppShell = React.memo(
       !!captionMessage &&
       !!workstationActiveSessionId;
 
-    // Seed the empty tab pool with a Start launcher tab on first mount
-    // (regular WorkStation only — Agent Station has its own surface).
-    useEnsureStartTab(!isAgentStation);
+    // Keep a Launchpad tab as the pool's home base: seed it when empty, drop
+    // it once real tabs exist (regular WorkStation only — Agent Station has
+    // its own surface).
+    useLaunchpadTab(!isAgentStation);
 
     const workStationPanels = useWorkStationPanels();
     useAppShellSimulatorPanelSync({ isAgentStation, workStationPanels });
