+++ src/modules/WorkStation/shared/TabBar/components/SortableTab/index.tsx
@@ -20,6 +20,7 @@ import {
   GitPullRequest,
   Globe,
   Layout,
+  LayoutGrid,
   LayoutList,
   ListChecks,
   Lock,
@@ -83,6 +84,7 @@ const WORKSTATION_TAB_ICONS = {
   Globe,
   Infinity,
   Layout,
+  LayoutGrid,
   LayoutList,
   ListChecks,
   MessageCircle,
